Correctly use Cstruct buffer offset
Cstruct buffers may be slices of larger buffers, in which case it is not
sufficient to just pass through the buffer and length; this adds an
additional argument to pass b.Cstruct.off down to the mirage-solo5
stubs. Bump mirage-solo5 dependency to >= 0.5.0 accordingly.

Note that Cstruct buffers with a non-zero off are not currently used
anywhere in the network stack; this is merely future-proofing while
doing the same changes for the block stubs where they are required.