Update to new Solo5 "manifest"-style APIs
This is a minimal update to use the new Solo5 APIs with support for
multiple devices.

As there is no integration with the Lwt main loop yet, this code will
NOT yet function correctly in the presence of multiple network devices
in a unikernel. Full functionality for multiple network device operation
will come in a subsequent commit.