Be able to opam install mirage-net-solo5
This package is not usable into the default context and should be
installed _via_ `opam install`. However, for the purpose of the
documentation, we want to provide it. This patch allows the package
to be installable (but not usable) _via_ `opam install` and take the
right C stub depending on the `dune`'s context.

Fortunately, `dune` will prioritize the `mirage-net-solo5` avaialble
into sources instead of the installed one. In other words, an
installation of mirage-net-solo5 can co-exist with sources pulled _via_
`opam monorepo`. When you will build an unikernel, `dune` will take
sources and the context will help us to choose the right C stub.